+++ electron/listeners/fileIndexies.ts
@@ -2,11 +2,7 @@ import type { BrowserWindow, IpcMainEvent } from 'electron'
 import { resolve } from 'path'
 
 import { copyAllHCSFilesAsync } from '../helpers/asyncCopy'
-import { builderTasks, findIndexiesFiles, IReadedIndex, loadIndexFile } from '../helpers/files'
-
-// const wait = (timeout: number): Promise<any> => {
-//   return new Promise(resolve => setTimeout(resolve, timeout))
-// }
+import { builderTasks, findIndexiesFiles, IReadedIndex, listFiles, loadIndexFile } from '../helpers/files'
 
 export function findIndexies(mainWindow: BrowserWindow, event: IpcMainEvent, path?: string) {
   const listFiles = path ? findIndexiesFiles(path) : []
@@ -14,9 +10,25 @@ export function findIndexies(mainWindow: BrowserWindow, event: IpcMainEvent, pat
 }
 
 export function loadIndexies(mainWindow: BrowserWindow, event: IpcMainEvent, paths: string[] = []) {
-  const indexFiles = paths.map(loadIndexFile)
+  const indexFiles = paths
+    .map(loadIndexFile)
+    .reduce((acc, items) => {
+      items.forEach(item => acc.push(item))
+      return acc
+    }, [])
+    .map((item, i, arr) => {
+      const report = resolve(item.inDir, item.dateDir, item.timeDir)
+      arr[i].files = listFiles(report)
+      event.sender.send('onReadProgress', { total: arr.length, current: i + 1 })
+      if (arr[i].files.length) {
+        event.sender.send('onReadDir', arr[i])
+        return arr[i]
+      }
+      return null
+    })
+    .filter(f => !!f?.files?.length)
+
   event.sender.send('loadedIndexies', indexFiles)
-  return indexFiles
 }
 
 export async function processAll(
@@ -33,4 +45,22 @@ export async function processAll(
     })
   )
   event.sender.send('processedAll', tasks)
+}
+
+export async function processFile(
+  mainWindow: BrowserWindow,
+  event: IpcMainEvent,
+  fileReaded: IReadedIndex,
+  outDir: string
+) {
+  const [task] = await Promise.all(
+    builderTasks([fileReaded]).map(async task => {
+      event.sender.send(`startProcessItem-${task.id}`)
+      const copied = await copyAllHCSFilesAsync(task.from, resolve(outDir, task.outDir))
+      const taskProcessed = { ...task, copied: copied.length }
+      event.sender.send(`processedItem-${task.id}`, taskProcessed)
+      return taskProcessed
+    })
+  )
+  event.sender.send(`processedFile-${task.id}`, task)
 }