+++ src/components/Organize/SideSummary/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useMemo } from 'react'
+import React, { useCallback, useState } from 'react'
 
 import { AppContainer } from '../../AppContainer'
 import { useAppTheme } from '../../AppTheme/useAppTheme'
@@ -8,50 +8,58 @@ import { useOrganize } from '../OrganizeProvider'
 export const SideSummary: React.FC = () => {
   const { theme } = useAppTheme()
   const [inputPath, setInputPath] = useState('')
-  const { findFileIndexies, loadedIndex, setOutPath } = useOrganize()
+  const { findFileIndexies, loadedIndex, loading, clearLoadedIndex, outPath, setOutPath } = useOrganize()
 
   const handleSelectInput = useCallback(async () => {
-    const [path] = await window.Main.sendOpenDir()
+    const [path] = await window.Main.sendOpenDir(inputPath)
     if (path) {
-      findFileIndexies(path)
       setInputPath(path)
-      setOutPath(old => (!old ? `${path}/output` : old))
+      clearLoadedIndex()
+      await findFileIndexies(path)
+      // setOutPath(old => (!old ? `${path}/output` : old))
     }
-  }, [findFileIndexies, setOutPath])
+  }, [findFileIndexies, clearLoadedIndex, inputPath])
 
-  const countReport = useMemo(() => {
-    return loadedIndex.reduce((acc, item) => {
-      const count = item.length
-      return acc + count
-    }, 0)
-  }, [loadedIndex])
+  const handleSelectOutPut = useCallback(async () => {
+    const [path] = await window.Main.sendOpenDir(outPath)
+    if (path) {
+      setOutPath(path)
+    }
+  }, [setOutPath, outPath])
 
   return (
     <AppContainer horizontalSpaced>
       <SimpleText verticalSpaced color={theme.colors.textDark} size={14}>
         Selecione a pasta raiz para localizar os arquivos de índice do SDCArd <em>(INDEX.TXT).</em>
       </SimpleText>
-      <ActionButton onClick={handleSelectInput}>{'SELECIONE'}</ActionButton>
+      <ActionButton disabled={!!loading} onClick={handleSelectInput}>
+        {'SELECIONE'}
+      </ActionButton>
       <br />
       {inputPath ? (
         <SimpleText size={14} verticalSpaced align="left">
-          <strong>Local:</strong> {inputPath}
+          <strong>Entrada:</strong> {inputPath}
         </SimpleText>
       ) : null}
       {loadedIndex.length ? (
         <div>
           <SimpleText size={14} verticalSpaced align="left">
-            <strong>Indices encontrados:</strong> {loadedIndex.length}
+            <strong>Relatórios:</strong> {loadedIndex.length}
           </SimpleText>
-        </div>
-      ) : null}
-      {countReport ? (
-        <div>
+          <hr />
+          <SimpleText verticalSpaced color={theme.colors.textDark} size={14}>
+            Selecione a pasta de saída para organizar os arquivos.
+          </SimpleText>
+          <ActionButton disabled={!!loading} onClick={handleSelectOutPut}>
+            {'SELECIONE'}
+          </ActionButton>
+          <br />
           <SimpleText size={14} verticalSpaced align="left">
-            <strong>Relatórios:</strong> {countReport}
+            <strong>Saída:</strong> {outPath}
           </SimpleText>
         </div>
       ) : null}
+
       {inputPath && !loadedIndex.length ? (
         <SimpleText size={14} verticalSpaced align="center">
           <strong>Nenhum arquivo de índice HCS encontrado</strong>