+++ CHANGELOG.md
@@ -2,4 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.0.2](https://github.com/leguass7/guarita-hcs-organize/compare/v1.0.1...v1.0.2) (2021-10-10)
+
+
+### Features
+
+* adicionado progresso da cópia ([7742d86](https://github.com/leguass7/guarita-hcs-organize/commit/7742d866f078e8828d01e37ada152db7508676bd))
+* adiconado loading circular progress ([6ba4759](https://github.com/leguass7/guarita-hcs-organize/commit/6ba4759134fa291dc990112334fa0476accabf31))
+
 ### 1.0.1 (2021-10-09)